Move charm library helper functions to their own modules.
Fixes CRAFT-1557
