+++ charmcraft/errors.py
@@ -1,0 +1,37 @@
+# Copyright 2023 Canonical Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# For further info, check https://github.com/canonical/charmcraft
+"""Charmcraft error classes."""
+
+from craft_cli import CraftError
+
+
+class BadLibraryPathError(CraftError):
+    """Subclass to provide a specific error for a bad library path."""
+
+    def __init__(self, path):
+        super().__init__(
+            "Charm library path {} must conform to lib/charms/<charm>/vN/<libname>.py"
+            "".format(path)
+        )
+
+
+class BadLibraryNameError(CraftError):
+    """Subclass to provide a specific error for a bad library name."""
+
+    def __init__(self, name):
+        super().__init__(
+            "Charm library name {!r} must conform to charms.<charm>.vN.<libname>".format(name)
+        )
